Created menu.py file and added button
Created a menu.py file that holds a PvP button function. Can be refactored or changed depending on the architecture we'd like to create. Button code can be copy and pasted and changed for other features. Currently does not function as a button, clicking does not start a game. Feel free to make any changes.